Move FilePicker into form directory
