--- src/components/FilePicker.scss
@@ -1,5 +1,0 @@
-.file-picker {
-  padding: 1rem;
-  background: darken($background-color, 5%);
-  border-radius: 0.1875rem;
-}
